Improve build system
The glossaries are created correctly without manual steps now.  This
wasn't the case when doing a clean build before, because the second
invocation of `latexmk` wouldn't do anything.  It probably was broken in
other ways as well.

The `latexmkrc` file is from <https://tex.stackexchange.com/a/44316> and
tells `latexmk` when and how to run `makeglossaries`.  Apparently it
will do an unnecessary run in some cases, though.  See
<https://tex.stackexchange.com/q/60204>